Support creating a MHSM pool.

+++ src/KeyVault/KeyVault/Commands/RemoveAzureKeyVault.cs
@@ -71,7 +71,7 @@ namespace Microsoft.Azure.Commands.KeyVault
             ValueFromPipeline = true,
             HelpMessage = "Key Vault object to be deleted.")]
         [ValidateNotNullOrEmpty]
-        public PSKeyVault InputObject { get; set; }
+        public PSKeyVaultIdentityItem InputObject { get; set; }
 
         /// <summary>
         /// Vault object
@@ -134,6 +134,10 @@ namespace Microsoft.Azure.Commands.KeyVault
             HelpMessage = "Remove the previously deleted vault permanently.")]
         public SwitchParameter InRemovedState { get; set; }
 
+        [Parameter(Mandatory = false,            
+            HelpMessage = "Specifies the type of this vault as MHSM.")]
+        public SwitchParameter Hsm { get; set; }
+
         /// <summary>
         /// If present, do not ask for confirmation
         /// </summary>
@@ -180,9 +184,16 @@ namespace Microsoft.Azure.Commands.KeyVault
                         VaultName,
                         () =>
                         {
-                            KeyVaultManagementClient.PurgeVault(
-                                vaultName: VaultName,
-                                location: Location);
+                            if (Hsm)
+                            {
+                                //PurgeManagedHsm();
+                            }
+                            else
+                            {
+                                KeyVaultManagementClient.PurgeVault(
+                                    vaultName: VaultName,
+                                    location: Location);
+                            }
 
                             if (PassThru)
                             {
@@ -192,7 +203,8 @@ namespace Microsoft.Azure.Commands.KeyVault
             }
             else
             {
-                ResourceGroupName = string.IsNullOrWhiteSpace(ResourceGroupName) ? GetResourceGroupName(VaultName) : ResourceGroupName;
+                // Get resource group name for keyvault or ManagedHsm
+                ResourceGroupName = string.IsNullOrWhiteSpace(ResourceGroupName) ? GetResourceGroupName(VaultName,Hsm.IsPresent) : ResourceGroupName;
                 if (string.IsNullOrWhiteSpace(ResourceGroupName))
                     throw new ArgumentException(string.Format(Resources.VaultNotFound, VaultName, ResourceGroupName));
                 ConfirmAction(
@@ -208,9 +220,16 @@ namespace Microsoft.Azure.Commands.KeyVault
                     VaultName,
                     () =>
                     {
-                        KeyVaultManagementClient.DeleteVault(
-                    vaultName: VaultName,
-                    resourceGroupName: this.ResourceGroupName);
+                        if (Hsm)
+                        {
+                            // DeleteManagedHsm();
+                        }
+                        else
+                        {
+                            KeyVaultManagementClient.DeleteVault(
+                                vaultName: VaultName,
+                                resourceGroupName: this.ResourceGroupName);
+                        }
 
                         if (PassThru)
                         {
